しょぼいけど selfcompile 向け修正

+++ util.c
@@ -20,7 +20,7 @@ noreturn void verror_at(char *loc, char *fmt, va_list ap) {
     }
 
     int indent = fprintf(stderr, "%s:%d: ", input_filename, line_num);
-    fprintf(stderr, "%.*s\n", (int)(end - line), line);
+    fprintf(stderr, "%.*s\n", (end - line), line);
 
     int pos = loc - line + indent;
     fprintf(stderr, "%*s", pos, " ");
